refactor: update DashboardContext imports and remove unused useMemoCharacter hook
