+++ src/components/dashboard/WorkList.tsx
@@ -4,7 +4,7 @@ import {
   EmptyListDiscription,
 } from "@/styles/dashboard/WorkList";
 import { useContext } from "react";
-import { DashboardContext } from "@/hooks/dashboard/dashboard";
+import { DashboardContext } from "@/hooks/dashboard/workStudioAndTrash";
 import WorkButton from "./WorkButton";
 import { workspaceCategory } from "@/utils/APIs/types";
 import { AddWork } from "./ColoredRoundButtons";
