+++ src/hooks/dashboard/ideaBox.ts
@@ -1,13 +1,20 @@
-import { useQuery, useQueryClient } from "@tanstack/react-query";
-import { dashboardQueryKeys } from "@/utils/APIs/queryKeys";
-import { createContext, use, useEffect, useState } from "react";
-import { ideaBoxCategory } from "@/utils/APIs/types";
-import { getMemo } from "@/utils/APIs/dashboard";
+import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { memoQueryKeys } from "@/utils/APIs/queryKeys";
+import { createContext, use, useCallback, useEffect, useState } from "react";
+import { ideaBoxCategory, TMemo } from "@/utils/APIs/types";
+import {
+  createMemo,
+  deleteMemo,
+  getMemoList,
+  updaetMemoName,
+  updateMemoDescription,
+} from "@/utils/APIs/memo";
+import { debounce } from "@/utils";
 
 export function useIdeaBox() {
   const queryClient = useQueryClient();
+  const [memoList, setMemoList] = useState<TMemo[]>([]);
 
-  // 아이디어 보관함 카테고리
   const [ideaCategory, setIdeaCategory] = useState<
     keyof typeof ideaBoxCategory
   >(() => {
@@ -19,19 +26,117 @@ export function useIdeaBox() {
     }
     return "memo";
   });
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      localStorage.setItem("ideaCategory", ideaCategory);
-    }
-  }, [ideaCategory]);
+  const { data, error, isLoading } = useQuery({
+    queryKey: memoQueryKeys.memoList(),
+    queryFn: getMemoList,
+  });
+  const {
+    mutate: createMemoMutation,
+    isPending: isCreating,
+    mutateAsync: createMemoAsync,
+  } = useMutation({
+    mutationFn: createMemo,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: memoQueryKeys.memoList() });
+    },
+  });
+  const { mutate: updateMemoNameMutation, isPending: isUpdatingName } =
+    useMutation({ mutationFn: updaetMemoName });
+  const {
+    mutate: updateMemoDescriptionMutation,
+    isPending: isUpdatingDescription,
+  } = useMutation({ mutationFn: updateMemoDescription });
+  const { mutate: deleteMemoMutation, isPending: isDeleting } = useMutation({
+    mutationFn: deleteMemo,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: memoQueryKeys.memoList() });
+    },
+    onMutate: async (id: string) => {
+      const previousMemoList = queryClient.getQueryData<TMemo[]>(
+        memoQueryKeys.memoList()
+      );
+      queryClient.setQueryData<TMemo[]>(
+        memoQueryKeys.memoList(),
+        (old) => old?.filter((memo) => memo.id !== id) || []
+      );
+      return { previousMemoList };
+    },
+  });
+
+  const debounceUpdateMemoName = useCallback(
+    debounce(updateMemoNameMutation, 500),
+    [data]
+  );
+  const debounceUpdateMemoDescription = useCallback(
+    debounce(updateMemoDescriptionMutation, 500),
+    [data]
+  );
+  const onClickCreateMemo = () => {
+    createMemoMutation();
+  };
+  const onClickDeleteMemo = (id: string) => () => {
+    if (!window.confirm("정말로 삭제하시겠습니까?")) return;
+    deleteMemoMutation(id);
+  };
+
+  const onChangeMemoName =
+    (id: string) => (e: React.ChangeEvent<HTMLInputElement>) => {
+      debounceUpdateMemoName({ id, memo_name: e.target.value });
+      setMemoList((old) =>
+        old.map((memo) =>
+          memo.id === id ? { ...memo, memo_name: e.target.value } : memo
+        )
+      );
+    };
+  const onChangeMemoDescription =
+    (id: string) => (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+      debounceUpdateMemoDescription({ id, memo_description: e.target.value });
+      setMemoList((old) =>
+        old.map((memo) =>
+          memo.id === id ? { ...memo, memo_description: e.target.value } : memo
+        )
+      );
+    };
+  const onChangeMemoStart = async (
+    event: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    const memoId = await createMemoAsync();
+    updateMemoDescriptionMutation({
+      id: memoId,
+      memo_description: event.target.value,
+    });
+  };
   function handleIdeaCategoryChange(
     category: (typeof ideaBoxCategory)[keyof typeof ideaBoxCategory]
   ) {
     setIdeaCategory(category);
   }
 
+  useEffect(() => {
+    if (typeof window !== "undefined") {
+      localStorage.setItem("ideaCategory", ideaCategory);
+    }
+  }, [ideaCategory]);
+  useEffect(() => {
+    if (data) {
+      setMemoList(data);
+    }
+  }, [data]);
+
   return {
     ideaCategory,
+    memoList,
+    error,
+    isLoading,
+    isCreating,
+    isUpdatingName,
+    isUpdatingDescription,
+    isDeleting,
+    onClickCreateMemo,
+    onClickDeleteMemo,
+    onChangeMemoName,
+    onChangeMemoDescription,
+    onChangeMemoStart,
     handleIdeaCategoryChange,
   };
 }
